Refactor messaging tests into own module
